renaming files and methods for consistency and clearity
